+++ src/main/java/com/jnape/palatable/lambda/adt/hlist/Tuple3.java
@@ -57,6 +57,16 @@ public class Tuple3<_1, _2, _3> extends HCons<_1, Tuple2<_2, _3>> implements
     }
 
     /**
+     * Snoc an element onto the back of this HList.
+     *
+     * @param _4   the new last element
+     * @return the updated HList
+     */
+    public <_4> Tuple4<_1, _2, _3, _4> snoc(_4 _4) {
+        return tuple(_1, _2, _3, _4);
+    }
+
+    /**
      * {@inheritDoc}
      */
     @Override